+++ Sources/DTTableViewManager/DTTableViewDelegate.swift
@@ -104,6 +104,9 @@ open class DTTableViewDelegate : DTTableViewDelegateWrapper, UITableViewDelegate
     
     /// Implementation for `UITableViewDelegate` protocol
     open func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        guard shouldDisplayHeaderView(forSection: section) else {
+            return configuration?.minimalHeaderHeightForTableView(tableView) ?? .zero
+        }
         if let height = performHeaderReaction(.heightForHeaderInSection, location: section, provideView: false) as? CGFloat {
             return height
         }
@@ -139,6 +142,9 @@ open class DTTableViewDelegate : DTTableViewDelegateWrapper, UITableViewDelegate
     
     /// Implementation for `UITableViewDelegate` protocol
     open func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
+        guard shouldDisplayFooterView(forSection: section) else {
+            return configuration?.minimalFooterHeightForTableView(tableView) ?? .zero
+        }
         if let height = performFooterReaction(.heightForFooterInSection, location: section, provideView: false) as? CGFloat {
             return height
         }